+++ GatewayServiceTest/WebContent/Component.js
@@ -9,6 +9,7 @@ sap.ui.define([
       init : function () {
          // call the init function of the parent
          UIComponent.prototype.init.apply(this, arguments);
+         this.getRouter().initialize();
       }
    });
-});';//////'
+});
