--- GatewayServiceTest/WebContent/WEB-INF/web.xml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-							http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-		version="2.5"
-		xmlns="http://java.sun.com/xml/ns/javaee">
-
-	<display-name>GatewayServiceTest</display-name>
-
-	<!-- ============================================================== -->
-	<!-- UI5 resource servlet used to handle application resources      -->
-	<!-- ============================================================== -->
-
-	<servlet>
-		<display-name>ResourceServlet</display-name>
-		<servlet-name>ResourceServlet</servlet-name>
-		<servlet-class>com.sap.ui5.resource.ResourceServlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>ResourceServlet</servlet-name>
-		<url-pattern>/resources/*</url-pattern>
-	</servlet-mapping>
-	<!-- BEGIN: test-resources -->
-	<!-- test-resources mapping has to be switched off for productive use on a Java server! -->
-	<servlet-mapping>
-		<servlet-name>ResourceServlet</servlet-name>
-		<url-pattern>/test-resources/*</url-pattern>
-	</servlet-mapping>
-	<!-- END: test-resources -->
-	<!-- BEGIN: DEV MODE -->
-	<!-- DEV MODE switched off by default and can be switched on during development -->
-	<!-- but has to be switched off for productive use on a Java server! -->
-	<context-param>
-		<param-name>com.sap.ui5.resource.DEV_MODE</param-name>
-		<param-value>false</param-value>
-	</context-param>
-	<!-- END: DEV MODE -->
-
-
-	<!-- ============================================================== -->
-	<!-- Cache Control Filter to prevent caching of any resource        -->
-	<!-- ============================================================== -->
-
-	<filter>
-		<display-name>CacheControlFilter</display-name>
-		<filter-name>CacheControlFilter</filter-name>
-		<filter-class>com.sap.ui5.resource.CacheControlFilter</filter-class>
-	</filter>
-	<filter-mapping>
-		<filter-name>CacheControlFilter</filter-name>
-		<url-pattern>*.html</url-pattern>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>CacheControlFilter</filter-name>
-		<url-pattern>*.js</url-pattern>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>CacheControlFilter</filter-name>
-		<url-pattern>*.xml</url-pattern>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>CacheControlFilter</filter-name>
-		<url-pattern>*.json</url-pattern>
-	</filter-mapping>
-	<filter-mapping>
-		<filter-name>CacheControlFilter</filter-name>
-		<url-pattern>*.css</url-pattern>
-	</filter-mapping>
-
-
-	<!-- ============================================================== -->
-	<!-- UI5 proxy servlet                                              -->
-	<!-- ============================================================== -->
-
-	<servlet>
-		<servlet-name>SimpleProxyServlet</servlet-name>
-		<servlet-class>com.sap.ui5.proxy.SimpleProxyServlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>SimpleProxyServlet</servlet-name>
-		<url-pattern>/proxy/*</url-pattern>
-	</servlet-mapping>
-
-
-	<!-- ============================================================== -->
-	<!-- Welcome file list                                              -->
-	<!-- ============================================================== -->
-
-	<welcome-file-list>
-		<welcome-file>index.html</welcome-file>
-	</welcome-file-list>
-
-
-</web-app>
