+++ GatewayServiceTest/WebContent/view/Home.view.xml
@@ -1,0 +1,60 @@
+<mvc:View controllerName="gatewayservicetest.controller.App" xmlns="sap.m" xmlns:mvc="sap.ui.core.mvc" displayBlock="true">
+	<App>
+		<pages>
+			<Page title="{i18n>homePageTitle}">
+				<content>
+					<Table id="list" growing="true" growingThreshold="10" title="{i18n>title}" items="{/SFLIGHT}" enableCellFilter="true" enableBusyIndicator="true">
+						<headerToolbar>
+							<OverflowToolbar>
+								<ToolbarSpacer/>
+								<SearchField id="carrid" liveChange="onFilter" width="10rem"/>
+								<Button text="Reset" type="Transparent" press="onReset"/>
+							</OverflowToolbar>
+						</headerToolbar>
+						<columns>
+							<Column>
+								<Text text="{i18n>carrid}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>connid}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>fldate}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>price}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>planetype}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>seatsmax}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>seatsocc}"/>
+							</Column>
+							<Column>
+								<Text text="{i18n>paymentsum}"/>
+							</Column>
+						</columns>
+						<items>
+							<ColumnListItem>
+								<cells>
+									<Text text="{Carrid}"/>
+									<Text text="{Connid}"/>
+									<Text text="{ path: 'Fldate', type: 'sap.ui.model.type.Date', formatOptions: { pattern: 'MM/dd/yyyy' } }"/>
+									<Text text="{ parts: [ {path: 'Price'}, {path: 'Currency'} ], type: 'sap.ui.model.type.Currency', formatOptions: { showMeasure: true } }"/>
+									<Text text="{Planetype}"/>
+									<Text text="{Seatsmax}"/>
+									<Text text="{Seatsocc}"/>
+									<Text
+										text="{ parts: [ {path: 'Paymentsum'}, {path: 'Currency'} ], type: 'sap.ui.model.type.Currency', formatOptions: { showMeasure: true } }"/>
+								</cells>
+							</ColumnListItem>
+						</items>
+					</Table>
+				</content>
+			</Page>
+		</pages>
+	</App>
+</mvc:View>
